dtos added

+++ src/main/java/com/anirudh/portfolio/aniapp/model/LanguageProficiency.java
@@ -11,7 +11,7 @@ public class LanguageProficiency {
     @SequenceGenerator(name = "lang_prof_generator", allocationSize = 5, sequenceName = "language_proficiency_sequence", schema = "portfolio")
     @GeneratedValue(generator = "lang_prof_generator", strategy = GenerationType.SEQUENCE)
     @Column(name = "LANGUAGE_ID")
-    private int languageId;
+    private long languageId;
     @Column(name = "LANGUAGE", nullable = false)
     private String name;
     @Column(name = "LANGUAGE_PROFICIENCY", columnDefinition = "INTEGER DEFAULT 1")
