+++ src/main/java/com/anirudh/portfolio/aniapp/model/Skill.java
@@ -11,7 +11,7 @@ public class Skill {
     @SequenceGenerator(name = "skill_generator", allocationSize = 5, sequenceName = "skill_sequence", schema = "portfolio")
     @GeneratedValue(generator = "skill_generator", strategy = GenerationType.SEQUENCE)
     @Column(name = "SKILL_ID")
-    private int skillId;
+    private long skillId;
     @Column(name = "SKILL_NAME", nullable = false)
     private String name;
     @Column(name = "SKILL_PROFICIENCY", columnDefinition = "INTEGER DEFAULT 1")
